feat: Implementa módulo Tema e relacionamentos

+++ src/postagem/services/postagem.service.ts
@@ -1,33 +1,35 @@
-import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
-import { DeleteResult, ILike, Repository } from 'typeorm';
-import { Postagem } from '../entities/postagem.entity';
-import { TemaService } from '../../tema/services/tema.service';
+import { HttpException, HttpStatus, Injectable } from "@nestjs/common";
+import { InjectRepository } from "@nestjs/typeorm";
+import { DeleteResult, ILike, Repository } from "typeorm";
+import { TemaService } from "../../tema/services/tema.service";
+import { Postagem } from "../entities/postagem.entity";
 
 @Injectable()
 export class PostagemService {
     constructor(
         @InjectRepository(Postagem)
         private postagemRepository: Repository<Postagem>,
-        private temaService: TemaService
+        private temaService:TemaService
     ) { }
 
     async findAll(): Promise<Postagem[]> {
         return await this.postagemRepository.find({
             relations:{
-                tema: true
+                tema: true,
+                usuario: true
             }
         });
     }
 
     async findById(id: number): Promise<Postagem> {
 
-        const postagem = await this.postagemRepository.findOne({
+        let postagem = await this.postagemRepository.findOne({
             where: {
-                id,
+                id
             },
             relations:{
-                tema: true
+                tema: true,
+                usuario: true
             }
         });
 
@@ -37,36 +39,61 @@ export class PostagemService {
         return postagem;
     }
 
-    async findAllByTitulo(titulo: string): Promise<Postagem[]> {
+    async findByTitulo(titulo: string): Promise<Postagem[]> {
         return await this.postagemRepository.find({
             where:{
                 titulo: ILike(`%${titulo}%`)
             },
             relations:{
-                tema: true
+                tema: true,
+                usuario: true
             }
         })
     }
 
     async create(postagem: Postagem): Promise<Postagem> {
+       
+        if (postagem.tema){
+            
+            let tema = await this.temaService.findById(postagem.tema.id)
 
-        await this.temaService.findById(postagem.tema.id);
+            if (!tema)
+                throw new HttpException('Tema não encontrado!', HttpStatus.NOT_FOUND);
+            
+            return await this.postagemRepository.save(postagem);
+
+        }
 
         return await this.postagemRepository.save(postagem);
     }
 
-    async update(id: number, postagem: Postagem): Promise<Postagem> {
+    async update(postagem: Postagem): Promise<Postagem> {
         
-        await this.findById(postagem.id);
+        let buscaPostagem: Postagem = await this.findById(postagem.id);
 
-        await this.temaService.findById(postagem.tema.id);
+        if (!buscaPostagem || !postagem.id)
+            throw new HttpException('Postagem não encontrada!', HttpStatus.NOT_FOUND);
+
+        if (postagem.tema){
+            
+            let tema = await this.temaService.findById(postagem.tema.id)
+                
+            if (!tema)
+                throw new HttpException('Tema não encontrado!', HttpStatus.NOT_FOUND);
+                
+            return await this.postagemRepository.save(postagem);
+    
+        }
         
-        return this.postagemRepository.save(postagem);
+        return await this.postagemRepository.save(postagem);
     }
 
     async delete(id: number): Promise<DeleteResult> {
         
-        await this.findById(id);
+        let buscaPostagem = await this.findById(id);
+
+        if (!buscaPostagem)
+            throw new HttpException('Postagem não encontrada!', HttpStatus.NOT_FOUND);
 
         return await this.postagemRepository.delete(id);
